Move SnapshotError to agave-snapshots
