+++ runtime/src/accounts_background_service.rs
@@ -13,8 +13,8 @@ use {
         bank_forks::BankForks,
         snapshot_controller::SnapshotController,
         snapshot_package::{SnapshotKind, SnapshotPackage},
-        snapshot_utils::SnapshotError,
     },
+    agave_snapshots::error::SnapshotError,
     crossbeam_channel::{Receiver, SendError, Sender},
     log::*,
     rayon::iter::{IntoParallelIterator, ParallelIterator},
