+++ runtime/src/serde_snapshot.rs
@@ -6,10 +6,11 @@ use {
         epoch_stakes::VersionedEpochStakes,
         rent_collector::RentCollector,
         runtime_config::RuntimeConfig,
-        snapshot_utils::{SnapshotError, StorageAndNextAccountsFileId},
+        snapshot_utils::StorageAndNextAccountsFileId,
         stake_account::StakeAccount,
         stakes::{serialize_stake_accounts_to_delegation_format, Stakes},
     },
+    agave_snapshots::error::SnapshotError,
     bincode::{self, config::Options, Error},
     log::*,
     serde::{de::DeserializeOwned, Deserialize, Serialize},
