+++ snapshots/Cargo.toml
@@ -21,11 +21,13 @@ agave-unstable-api = []
 
 [dependencies]
 agave-fs = { workspace = true }
+bincode = { workspace = true }
 bzip2 = { workspace = true }
 crossbeam-channel = { workspace = true }
 log = { workspace = true }
 lz4 = { workspace = true }
 rand = { workspace = true }
+semver = { workspace = true }
 solana-accounts-db = { workspace = true }
 solana-clock = { workspace = true }
 solana-genesis-config = { workspace = true }
